Use ESGetToekn in some of the codes in SimTracker/Track(Vertex)Association
