+++ SimTracker/TrackAssociation/plugins/MCTrackMatcher.cc
@@ -36,7 +36,6 @@ private:
 };
 
 #include "DataFormats/Common/interface/Handle.h"
-#include "FWCore/Framework/interface/ESHandle.h"
 #include "FWCore/Framework/interface/Event.h"
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
 #include "FWCore/Utilities/interface/EDMException.h"
