+++ SimTracker/TrackAssociation/plugins/TrackAssociatorEDProducer.cc
@@ -14,8 +14,6 @@
 #include "FWCore/Framework/interface/Event.h"
 #include "FWCore/Framework/interface/MakerMacros.h"
 
-#include "FWCore/Framework/interface/ESHandle.h"
-
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
 
 #include "SimDataFormats/Associations/interface/TrackToTrackingParticleAssociator.h"
