+++ SimTracker/TrackAssociation/plugins/TrackMCQuality.cc
@@ -27,7 +27,6 @@
 #include "FWCore/Framework/interface/Event.h"
 #include "FWCore/Framework/interface/MakerMacros.h"
 
-#include "FWCore/Framework/interface/ESHandle.h"
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
 
 #include "SimDataFormats/Associations/interface/TrackToTrackingParticleAssociator.h"
